Correction Type of atribbutes at methods in entities

+++ src/main/java/com/films/api/films_api/entities/User.java
@@ -17,7 +17,7 @@ public class User implements Serializable{
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     private String name;
 
@@ -34,7 +34,7 @@ public class User implements Serializable{
         
     }
 
-    public User(Integer id, String name, String email, String password, LocalDate bornDate, String userImg) {
+    public User(Long id, String name, String email, String password, LocalDate bornDate, String userImg) {
         this.id = id;
         this.name = name;
         this.email = email;
@@ -43,11 +43,11 @@ public class User implements Serializable{
         this.userImg = userImg;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
